// src/app/appmanager/model/implementationParameters.ts
/**
 * Kathra Applications Manager
 * Kathra Applications Manager
 *
 * OpenAPI spec version: 1.2.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiVersion } from './apiVersion';


export interface ImplementationParameters { 
    /**
     * Implementation's name
     */
    name: string;
    apiVersion: ApiVersion;
    /**
     * The language to use for source code generation
     */
    language: string;
    /**
     * Implementation's description
     */
    desc?: string;
}